IMPLEMENTATIONS: Service-layer. TESTING: some services and mappers.

+++ src/main/java/co/edu/unimagdalena/busreserve/domine/repositories/RouteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RouteRepository extends JpaRepository<Route, Integer> {
+public interface RouteRepository extends JpaRepository<Route, Long> {
     List<Route> findByOriginAndDestination(String origin, String destination);
     List<Route> findAllByOrderByCodeAsc();
 }
